refactor(billing): centralize plans into @chm/pricing + AI message model
Single source of truth for cloud pricing shared by dashboard and landing:

- New zero-dep @chm/pricing package holds canonical BILLING_PLANS + display
  helpers (planAiUsage/planRetention/...). Dashboard plans.ts re-exports it;
  landing data/pricing.ts derives its cards + comparison matrix from it, so the
  hand-synced mirror (and its drift risk) is gone.
- AI usage model: Free 5 msg/day (hard cap), Pro 100/day, Max 1000/day, each
  paid tier +$5 / 2000 overage. checkAiDailyLimit soft-caps paid tiers (never
  blocks on unbuilt overage metering); Free still hard-blocks.
- Relabel retention to 'Conversation & insights history' (AI chat + insights,
  not metrics).
- Pricing page: centered intro, reformatted self-host banner, yearly toggle
  shows struck monthly price vs discounted per-month.
