+++ apps/landing/astro.config.mjs
@@ -1,7 +1,21 @@
-import { defineConfig } from 'astro/config'
+import { fileURLToPath } from 'node:url'
 import sitemap from '@astrojs/sitemap'
+import { defineConfig } from 'astro/config'
 
 export default defineConfig({
   site: 'https://chmonitor.dev',
   integrations: [sitemap()],
+  vite: {
+    resolve: {
+      alias: {
+        // Shared single source of truth for pricing (also used by the dashboard).
+        // Zero-dep pure-data package, so importing it drags no app deps in.
+        '@chm/pricing': fileURLToPath(
+          new URL('../../packages/pricing/src/index.ts', import.meta.url)
+        ),
+      },
+    },
+    // The package source lives outside the app root; allow Vite to read it.
+    server: { fs: { allow: ['../..'] } },
+  },
 })
