+++ apps/landing/tsconfig.json
@@ -1,1 +1,9 @@
-{ "extends": "astro/tsconfigs/strict" }
+{
+  "extends": "astro/tsconfigs/strict",
+  "compilerOptions": {
+    "baseUrl": ".",
+    "paths": {
+      "@chm/pricing": ["../../packages/pricing/src/index.ts"]
+    }
+  }
+}
