+++ packages/pricing/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "@chm/pricing",
+  "version": "0.1.0",
+  "private": false,
+  "type": "module",
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "devDependencies": {
+    "typescript": "^6.0.0"
+  },
+  "files": [
+    "src",
+    "tsconfig.json"
+  ],
+  "license": "MIT",
+  "publishConfig": {
+    "access": "public"
+  }
+}
